Add supported devices section
If any one reads this:

There are AT&T and Sprint variants (NA) and Asia models that use a Qualcomm modem. They are completely different, and AFAIK lack updated blobs for later versions. Not debugging this remotely. No one has shown interest in these devices in the past 5 years, so I doubt this will change.

Other variants use different NFC chips. It could be possible to support these models, but I am not as I have better things to do.